Reorganized Scene Prefab system
It is now easier to to make copies of all assets belonging to a scene by
pressing ctrl+D on the scene prefab. E.g. Higlight the scene1 prefab
folder, then press ctrl+D. This copies the scene1 prefab folder to a new
scene2 prefab folder. This folder can then be dragged into a new scene
and to set up everything needed in a new scene.